+++ selene/schema/audits/mutations.py
@@ -26,6 +26,7 @@ from selene.schema.entities import (
 )
 
 from selene.schema.utils import (
+    RESET_UUID,
     get_gmp,
     require_authentication,
     get_text_from_element,
@@ -129,22 +130,11 @@ class CreateAuditInput(graphene.InputObjectType):
         description="Whether the audit should be alterable"
     )
     comment = graphene.String(description="Audit comment")
-    observers = graphene.List(
-        graphene.UUID,
-        description="List of UUIDs for users which should be allowed to "
-        "observe the audit",
-    )
     preferences = graphene.Field(
         AuditPreferencesInput, description="Preferences to set for the audit"
     )
     schedule_id = graphene.UUID(
         description="UUID of a schedule when the audit should be run"
-    )
-    schedule_periods = graphene.Int(
-        description=(
-            "A limit to the number of times the "
-            "audit will be scheduled, or 0 for no limit"
-        )
     )
 
 
@@ -166,17 +156,12 @@ class CreateAudit(graphene.Mutation):
 
         name = input_object.name
         alterable = input_object.alterable
-        schedule_periods = input_object.schedule_periods
         comment = input_object.comment
 
         if input_object.alert_ids is not None:
             alert_ids = [str(alert_id) for alert_id in input_object.alert_ids]
         else:
             alert_ids = None
-        if input_object.observers is not None:
-            observers = [str(observer) for observer in input_object.observers]
-        else:
-            observers = None
 
         schedule_id = (
             str(input_object.schedule_id)
@@ -247,8 +232,6 @@ class CreateAudit(graphene.Mutation):
             comment=comment,
             alert_ids=alert_ids,
             schedule_id=schedule_id,
-            schedule_periods=schedule_periods,
-            observers=observers,
             preferences=preferences,
         )
         return CreateAudit(audit_id=resp.get('id'))
@@ -258,32 +241,26 @@ class ModifyAuditInput(graphene.InputObjectType):
     """Input ObjectType for modifying an audit"""
 
     audit_id = graphene.UUID(
-        required=True, description="UUID of the audit to modify.", name='id'
+        description="UUID of the audit to modify.", name='id', required=True
     )
-    name = graphene.String(description="Audit name")
+    name = graphene.String(description="Audit name", required=True)
+    target_id = graphene.UUID(description="UUID of target", required=True)
     policy_id = graphene.UUID(
-        description=("UUID of policy. OpenVAS Default scanners only")
+        description=("UUID of policy. OpenVAS Default scanners only"),
+        required=True,
     )
-    target_id = graphene.UUID(description="UUID of target")
-    scanner_id = graphene.UUID(description="UUID of scanner")
+    scanner_id = graphene.UUID(description="UUID of scanner", required=True)
 
     alert_ids = graphene.List(
         graphene.UUID, description="List of UUIDs for alerts"
     )
     alterable = graphene.Boolean(description="Whether the audit is alterable")
     comment = graphene.String(description="Audit comment")
-    observers = graphene.List(graphene.String)
     preferences = graphene.Field(
         AuditPreferencesInput, description="Preferences to set for the audit"
     )
     schedule_id = graphene.UUID(
         description="UUID of a schedule when the audit should be run."
-    )
-    schedule_periods = graphene.Int(
-        description=(
-            "A limit to the number of times the "
-            "audit will be scheduled, or 0 for no limit."
-        )
     )
 
 
@@ -308,38 +285,21 @@ class ModifyAudit(graphene.Mutation):
         audit_id = str(input_object.audit_id)
         name = input_object.name
         comment = input_object.comment
-        schedule_periods = input_object.schedule_periods
         alterable = input_object.alterable
 
         if input_object.alert_ids is not None:
             alert_ids = [str(alert_id) for alert_id in input_object.alert_ids]
         else:
-            alert_ids = None
+            alert_ids = []
 
-        if input_object.observers is not None:
-            observers = [str(observer) for observer in input_object.observers]
-        else:
-            observers = None
+        target_id = str(input_object.target_id)
+        scanner_id = str(input_object.scanner_id)
+        policy_id = str(input_object.policy_id)
 
         schedule_id = (
             str(input_object.schedule_id)
             if input_object.schedule_id is not None
-            else None
-        )
-        scanner_id = (
-            str(input_object.scanner_id)
-            if input_object.scanner_id is not None
-            else None
-        )
-        target_id = (
-            str(input_object.target_id)
-            if input_object.target_id is not None
-            else None
-        )
-        policy_id = (
-            str(input_object.policy_id)
-            if input_object.policy_id is not None
-            else None
+            else RESET_UUID
         )
 
         preferences = {}
@@ -388,10 +348,8 @@ class ModifyAudit(graphene.Mutation):
             scanner_id=scanner_id,
             alterable=alterable,
             schedule_id=schedule_id,
-            schedule_periods=schedule_periods,
             comment=comment,
             alert_ids=alert_ids,
-            observers=observers,
             preferences=preferences,
         )