+++ selene/schema/tasks/mutations.py
@@ -26,6 +26,7 @@ from selene.schema.entities import (
 )
 
 from selene.schema.utils import (
+    RESET_UUID,
     get_gmp,
     require_authentication,
     get_text_from_element,
@@ -153,25 +154,14 @@ class CreateTaskInput(graphene.InputObjectType):
         description="List of UUIDs for alerts to be used for the task",
     )
     alterable = graphene.Boolean(
-        description="Whether the task should be alterable"
+        description="Whether the task should be alterable",
     )
     comment = graphene.String(description="Task comment")
-    observers = graphene.List(
-        graphene.String,
-        description="List of UUIDs for users which should be allowed to "
-        "observe the task",
-    )
     preferences = graphene.Field(
         TaskPreferencesInput, description="Preferences to set for the task"
     )
     schedule_id = graphene.UUID(
         description="UUID of a schedule when the task should be run"
-    )
-    schedule_periods = graphene.Int(
-        description=(
-            "A limit to the number of times the "
-            "task will be scheduled, or 0 for no limit"
-        )
     )
 
 
@@ -193,17 +183,13 @@ class CreateTask(graphene.Mutation):
 
         name = input_object.name
         alterable = input_object.alterable
-        schedule_periods = input_object.schedule_periods
         comment = input_object.comment
 
         if input_object.alert_ids is not None:
             alert_ids = [str(alert_id) for alert_id in input_object.alert_ids]
         else:
             alert_ids = None
-        if input_object.observers is not None:
-            observers = [str(observer) for observer in input_object.observers]
-        else:
-            observers = None
+
         schedule_id = (
             str(input_object.schedule_id)
             if input_object.schedule_id is not None
@@ -274,8 +260,6 @@ class CreateTask(graphene.Mutation):
             comment=comment,
             alert_ids=alert_ids,
             schedule_id=schedule_id,
-            schedule_periods=schedule_periods,
-            observers=observers,
             preferences=preferences,
         )
         return CreateTask(task_id=resp.get('id'))
@@ -285,42 +269,37 @@ class ModifyTaskInput(graphene.InputObjectType):
     """Input ObjectType for modifying a task"""
 
     task_id = graphene.UUID(
-        required=True, description="UUID of task to modify.", name='id'
+        description="UUID of task to modify", name='id', required=True
     )
-    name = graphene.String(description="Task name")
+    name = graphene.String(description="Task name", required=True)
+    target_id = graphene.UUID(
+        description="UUID of the target to be used", required=True
+    )
+    scanner_id = graphene.UUID(
+        description="UUID of the scanner to be used", required=True
+    )
+
     scan_config_id = graphene.UUID(
         description=(
             "UUID of the scan config to use for the scanner. "
             "Only for OpenVAS scanners"
         ),
+        required=True,
     )
-    target_id = graphene.UUID(description="UUID of the target to be used")
-    scanner_id = graphene.UUID(description="UUID of the scanner to be used")
 
     alert_ids = graphene.List(
         graphene.UUID,
         description="List of UUIDs for alerts to be used for the task",
     )
     alterable = graphene.Boolean(
-        description="Whether the task should be alterable"
+        description="Whether the task should be alterable",
     )
     comment = graphene.String(description="Task comment")
-    observers = graphene.List(
-        graphene.String,
-        description="List of UUIDs for users which should be allowed to "
-        "observe the task",
-    )
     preferences = graphene.Field(
         TaskPreferencesInput, description="Preferences to set for the task"
     )
     schedule_id = graphene.UUID(
         description="UUID of a schedule when the task should be run"
-    )
-    schedule_periods = graphene.Int(
-        description=(
-            "A limit to the number of times the "
-            "task will be scheduled, or 0 for no limit."
-        )
     )
 
 
@@ -345,38 +324,21 @@ class ModifyTask(graphene.Mutation):
         task_id = str(input_object.task_id)
         name = input_object.name
         comment = input_object.comment
-        schedule_periods = input_object.schedule_periods
         alterable = input_object.alterable
 
         if input_object.alert_ids is not None:
             alert_ids = [str(alert_id) for alert_id in input_object.alert_ids]
         else:
-            alert_ids = None
+            alert_ids = []
 
-        if input_object.observers is not None:
-            observers = [str(observer) for observer in input_object.observers]
-        else:
-            observers = None
+        target_id = str(input_object.target_id)
+        scanner_id = str(input_object.scanner_id)
+        config_id = str(input_object.scan_config_id)
 
         schedule_id = (
             str(input_object.schedule_id)
             if input_object.schedule_id is not None
-            else None
-        )
-        scanner_id = (
-            str(input_object.scanner_id)
-            if input_object.scanner_id is not None
-            else None
-        )
-        target_id = (
-            str(input_object.target_id)
-            if input_object.target_id is not None
-            else None
-        )
-        config_id = (
-            str(input_object.scan_config_id)
-            if input_object.scan_config_id is not None
-            else None
+            else RESET_UUID
         )
 
         preferences = {}
@@ -427,10 +389,8 @@ class ModifyTask(graphene.Mutation):
             scanner_id=scanner_id,
             alterable=alterable,
             schedule_id=schedule_id,
-            schedule_periods=schedule_periods,
             comment=comment,
             alert_ids=alert_ids,
-            observers=observers,
             preferences=preferences,
         )