+++ selene/tests/audits/test_create_audit.py
@@ -98,7 +98,6 @@ class CreateAuditTestCase(SeleneTestCase):
             alert_ids=None,
             alterable=None,
             comment=None,
-            observers=None,
             preferences={
                 'auto_delete': 'keep',
                 'auto_delete_data': 4,
@@ -108,7 +107,6 @@ class CreateAuditTestCase(SeleneTestCase):
                 'assets_apply_overrides': 'no',
             },
             schedule_id=None,
-            schedule_periods=None,
         )
 
     def test_create_audit_auto_delete_reports_none(
@@ -165,7 +163,6 @@ class CreateAuditTestCase(SeleneTestCase):
             alert_ids=None,
             alterable=None,
             comment=None,
-            observers=None,
             preferences={
                 'auto_delete': 'no',
                 'max_checks': 7,
@@ -174,5 +171,4 @@ class CreateAuditTestCase(SeleneTestCase):
                 'assets_apply_overrides': 'no',
             },
             schedule_id=None,
-            schedule_periods=None,
         )